Rename EphemeralCounterSource.ts file to CounterSource.ts
